add utility file for wallet functionalities
